Add opt for sequentializing policy_channel updates
There is a known issue of parallelized policy_channel updates failing
when separate requests try to mutate the same channel. New Relic has
confirmed this happens because an update to the policy_channel link is
actually a mutation of the channel itself which includes a list of
policies. That list is not locked, so parallelized requests stomp
down on each other, leading to errors when using infrastructure
management tools, like Terraform, that fire requests in close
succession.